+++ docs/source/changelog.rst
@@ -17,6 +17,19 @@ Changelog
 .. +++++++++
 
 
+v0.21.0 / 2021-11-22
+--------------------
+
+Enhancements
+++++++++++++
+- (:pr:`321`) CFOUR, GAMESS, NWChem, Psi4, DFTD3, MP2D, gCP - learned to return certain native text
+  files under control of the ``native_files`` protocol. GAMESS users are strongly advised to at
+  least set ``protocols.native_files = "input"`` so that the job is reproducible. @loriab
+- (:pr:`325`) Torsiondrive - learned to use multiple molecules as input to torsiondrives. @jthorton
+- (:pr:`327`) TorchANI - learned to use GPUs if available. @kexul
+- (:pr:`330`, :pr:`332`) NWChem - learned to restart from existing scratch if QCEngine is killed. @WardLT
+
+
 v0.20.1 / 2021-10-08
 --------------------
 
@@ -24,10 +37,11 @@ Bug Fixes
 +++++++++
 
 - (:pr:`322`) Psi4 - allowed more test cases with gradients and Hessians after a compatibility PR started
-  saving them.
+  saving them. @loriab
 - (:pr:`323`) Turbomole - learned to store calcinfo_natom so that gradients and Hessians can be computed
   after QCElemental started using that quantity for shape checking in 
   [v0.22.0](https://github.com/MolSSI/QCElemental/blob/master/docs/source/changelog.rst#0220--2021-08-26)
+  @eljost
 
 
 v0.20.0 / 2021-10-01
@@ -59,11 +73,11 @@ Enhancements
   was previously off by a little. @loriab
 - (:pr:`316`) NWChem - harness learned to obey scratch_messy and memory local_config options. Memory
   was previously very off for v7. @loriab
-- (:pr:`317`) CFOUR, GAMESS, NWChem -- learned to return in AtomicInput or program native orientation
+- (:pr:`315`) CFOUR, GAMESS, NWChem -- learned to return in AtomicInput or program native orientation
   depending on fix_com & fix_orientation= T or F. Psi4 already did this. Previously these three
   always returned AtomicInput orientation. Note that when returning program native orientation, the
   molecule is overwritten, so AtomicResult is not a superset of AtomicInput. @loriab
-- (:pr:`317`) CFOUR, GAMESS, NWChem -- learned to harvest gradients and Hessians. @loriab
+- (:pr:`315`) CFOUR, GAMESS, NWChem -- learned to harvest gradients and Hessians. @loriab
 - (:pr:`317`) Docs - start "new harness" docs, so contributors have a coarse roadmap. @loriab
 - (:pr:`318`) Docs - documentation is now served from https://molssi.github.io/QCEngine/ and built
   by https://github.com/MolSSI/QCEngine/blob/master/.github/workflows/CI.yml .